make main driver empty
